// frontend/client/src/types.tsx
export interface ChatMsg {
  id: number;
  userId: number;
  username: string;
  createdAt: string;
  message: string;
  likedByUserId: number[];
  userUUID: string;
}